210427-02

+++ remc2/sub_main.cpp
@@ -9047,8 +9047,8 @@ void sub_11960(type_event_0x6E8E* a1)//1f2960
 	v4 = a1->axis_0x4C_76.y;
 	while (v2x > x_DWORD_EA3E4[0])
 	{
-		if (abs((signed __int16)(v2x->axis_0x4C_76.x - v1)) <= v2x->array_0x52_82.xshift + v6
-			&& abs((signed __int16)(v2x->axis_0x4C_76.y - v4)) <= v2x->array_0x52_82.yshift + v5)
+		if (abs((signed __int16)((int16_t)v2x->axis_0x4C_76.x - v1)) <= v2x->array_0x52_82.xshift + v6
+			&& abs((signed __int16)((int16_t)v2x->axis_0x4C_76.y - v4)) <= v2x->array_0x52_82.yshift + v5)
 		{
 			v2x->dword_0x8 = -1;
 			v2x->byte_0x3D_61 = 0;
@@ -9216,8 +9216,8 @@ bool sub_11CB0(axis_3d* a1x)//1f2cb0
 	while (!v10 && v2x > x_DWORD_EA3E4[0])
 	{
 		if (v2x->subtype_0x40_64 == 2
-			&& abs((signed __int16)(v2x->axis_0x4C_76.x - v8)) <= v2x->array_0x52_82.xshift + 2560
-			&& abs((signed __int16)(v2x->axis_0x4C_76.y - v1)) <= v2x->array_0x52_82.yshift + 2560)
+			&& abs((signed __int16)((int16_t)v2x->axis_0x4C_76.x - v8)) <= v2x->array_0x52_82.xshift + 2560
+			&& abs((signed __int16)((int16_t)v2x->axis_0x4C_76.y - v1)) <= v2x->array_0x52_82.yshift + 2560)
 		{
 			v10 = 1;
 		}
@@ -9227,8 +9227,8 @@ bool sub_11CB0(axis_3d* a1x)//1f2cb0
 	{
 		for (ix = x_D41A0_BYTEARRAY_4_struct.dword_38527; !v10 && ix > x_DWORD_EA3E4[0]; ix = ix->next_0)
 		{
-			if (abs((signed __int16)(ix->axis_0x4C_76.x - v8)) <= ix->array_0x52_82.xshift + 2560
-				&& abs((signed __int16)(ix->axis_0x4C_76.y - v1)) <= ix->array_0x52_82.yshift + 2560)
+			if (abs((signed __int16)((int16_t)ix->axis_0x4C_76.x - v8)) <= ix->array_0x52_82.xshift + 2560
+				&& abs((signed __int16)((int16_t)ix->axis_0x4C_76.y - v1)) <= ix->array_0x52_82.yshift + 2560)
 			{
 				v10 = 1;
 			}
@@ -9778,10 +9778,10 @@ void sub_12500(type_event_0x6E8E* a1x)//1f3500
 			switch (D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].index_0x3647A_0)
 			{
 			case 1:
-				v1x = abs(D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.x - a1x->axis_0x4C_76.x);
+				v1x = abs((int16_t)D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.x - (int16_t)a1x->axis_0x4C_76.x);
 				if (v1x <= 2048)
 				{
-					v1x = abs(D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.y - a1x->axis_0x4C_76.y);
+					v1x = abs((int16_t)D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.y - (int16_t)a1x->axis_0x4C_76.y);
 					if (v1x <= 2048)
 						goto LABEL_44;
 				}
@@ -9822,10 +9822,10 @@ void sub_12500(type_event_0x6E8E* a1x)//1f3500
 						{
 							if (D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.x || D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.y)
 							{
-								v1x = abs(D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.x - a1x->axis_0x4C_76.x);
+								v1x = abs(D41A0_0.StageVars2_0x365F4[(int16_t)a1x->StageVar1_0x48_72].str_0x3647C_4.axis.x - (int16_t)a1x->axis_0x4C_76.x);
 								if (v1x <= 3072)
 								{
-									v1x = abs(D41A0_0.StageVars2_0x365F4[a1x->StageVar1_0x48_72].str_0x3647C_4.axis.y - a1x->axis_0x4C_76.y);
+									v1x = abs(D41A0_0.StageVars2_0x365F4[(int16_t)a1x->StageVar1_0x48_72].str_0x3647C_4.axis.y - (int16_t)a1x->axis_0x4C_76.y);
 									if (v1x <= 3072)
 										goto LABEL_44;
 								}
@@ -16604,9 +16604,9 @@ void sub_1BD90(type_event_0x6E8E* a1x, char a2)//1fcd90
 			{
 				if (!jx->word_0x32_50 && jx != a1x)
 				{
-					v10 = (signed __int16)(jx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
-					v11 = (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-						* (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
+					v10 = (signed __int16)((int16_t)jx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
+					v11 = (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+						* (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
 						+ v10 * v10;
 					if (v11 <= v15)
 					{
@@ -16724,9 +16724,9 @@ void sub_1BF90(type_event_0x6E8E* a1x, char a2)//1fcf90
 					v27 = v11 * v11;
 					while (v12x > x_DWORD_EA3E4[0])
 					{
-						v13 = (signed __int16)(v12x->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
-						v14 = (signed __int16)(v12x->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-							* (signed __int16)(v12x->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
+						v13 = (signed __int16)((int16_t)v12x->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
+						v14 = (signed __int16)((int16_t)v12x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+							* (signed __int16)((int16_t)v12x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
 							+ v13 * v13;
 						if (v14 <= v27 && !(v12x->struct_byte_0xc_12_15.byte[0] & 0x20))
 						{
@@ -16760,9 +16760,9 @@ void sub_1BF90(type_event_0x6E8E* a1x, char a2)//1fcf90
 							{
 								if (!jx->word_0x32_50 && jx != a1x)
 								{
-									v19 = (signed __int16)(jx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
-									v20 = (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-										* (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
+									v19 = (signed __int16)((int16_t)jx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
+									v20 = (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+										* (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
 										+ v19 * v19;
 									if (v20 <= v26)
 									{
@@ -17235,8 +17235,8 @@ void sub_1C980(type_event_0x6E8E* a1x, char a2)//1fd980
 				for (jx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[a1x->subtype_0x40_64]; jx > x_DWORD_EA3E4[0]; jx = jx->next_0)
 				{
 					if (jx->word_0x1A_26 != a1x->word_0x1A_26
-						&& abs(a1x->axis_0x4C_76.x - jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
-						&& abs(a1x->axis_0x4C_76.y - jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
+						&& abs((int16_t)a1x->axis_0x4C_76.x - (int16_t)jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
+						&& abs((int16_t)a1x->axis_0x4C_76.y - (int16_t)jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
 					{
 						a1x->word_0x20_32 = sub_581E0_maybe_tan2(&jx->axis_0x4C_76, &a1x->axis_0x4C_76);
 						break;
@@ -17794,16 +17794,16 @@ void sub_1D8C0(type_event_0x6E8E* a1x, __int16 a2)//1fe8c0
 					for (jx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[a1x->subtype_0x40_64]; jx > x_DWORD_EA3E4[0]; jx = jx->next_0)
 					{
 						if (jx->word_0x1A_26 != a1x->word_0x1A_26
-							&& abs(a1x->axis_0x4C_76.x - jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
-							&& abs(a1x->axis_0x4C_76.y - jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
+							&& abs((int16_t)a1x->axis_0x4C_76.x - (int16_t)jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
+							&& abs((int16_t)a1x->axis_0x4C_76.y - (int16_t)jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
 						{
 							a1x->word_0x20_32 = sub_581E0_maybe_tan2(&jx->axis_0x4C_76, &a1x->axis_0x4C_76);
 							break;
 						}
 					}
-					if (abs(a1x->axis_0x4C_76.x - v12x->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
+					if (abs((int16_t)a1x->axis_0x4C_76.x - (int16_t)v12x->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
 						+ v12x->array_0x52_82.xshift
-						&& abs(a1x->axis_0x4C_76.y - v12x->axis_0x4C_76.y) < a1x->array_0x52_82.xshift
+						&& abs((int16_t)a1x->axis_0x4C_76.y - (int16_t)v12x->axis_0x4C_76.y) < a1x->array_0x52_82.xshift
 						+ v12x->array_0x52_82.xshift)
 					{
 						a1x->word_0x20_32 = sub_581E0_maybe_tan2(&v12x->axis_0x4C_76, &a1x->axis_0x4C_76);
@@ -17881,8 +17881,8 @@ void sub_1DBF0(type_event_0x6E8E* a1x, unsigned __int16 a2)//1febf0
 				v11 = a1x->dword_0xA0_160x->word_160_0x1c_28 * a1x->dword_0xA0_160x->word_160_0x1c_28;
 				while (v6x > x_DWORD_EA3E4[0])
 				{
-					v7 = v6x->axis_0x4C_76.x - a1x->axis_0x4C_76.x;
-					v8 = v6x->axis_0x4C_76.y - a1x->axis_0x4C_76.y;
+					v7 = (int16_t)v6x->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x;
+					v8 = (int16_t)v6x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y;
 					v13 = v8 * v8 + v7 * v7;
 					//result = v13;
 					if (v13 <= v11 && !(v6x->struct_byte_0xc_12_15.byte[0] & 0x20))
@@ -18424,8 +18424,8 @@ void sub_1E700(type_event_0x6E8E* a1x, char a2)//1ff700
 					for (jx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[a1x->subtype_0x40_64]; jx > x_DWORD_EA3E4[0]; jx = jx->next_0)
 					{
 						if (jx->word_0x1A_26 != a1x->word_0x1A_26
-							&& abs(a1x->axis_0x4C_76.x - jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
-							&& abs(a1x->axis_0x4C_76.y - jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
+							&& abs((int16_t)a1x->axis_0x4C_76.x - (int16_t)jx->axis_0x4C_76.x) < a1x->array_0x52_82.xshift
+							&& abs((int16_t)a1x->axis_0x4C_76.y - (int16_t)jx->axis_0x4C_76.y) < a1x->array_0x52_82.xshift)
 						{
 							a1x->word_0x20_32 = sub_581E0_maybe_tan2(&jx->axis_0x4C_76, &a1x->axis_0x4C_76);
 							return;
@@ -19366,9 +19366,9 @@ void sub_1FAA0(type_event_0x6E8E* a1x)//200aa0
 				v31 = a1x->dword_0xA0_160x->word_160_0x1c_28 * a1x->dword_0xA0_160x->word_160_0x1c_28;
 				while (v10x > x_DWORD_EA3E4[0])
 				{
-					v12 = (signed __int16)(v10x->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
-					v35 = (signed __int16)(v10x->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-						* (signed __int16)(v10x->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
+					v12 = (signed __int16)((int16_t)v10x->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
+					v35 = (signed __int16)((int16_t)v10x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+						* (signed __int16)((int16_t)v10x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
 						+ v12 * v12;
 					//v35 = jx;
 					if (v35 <= v31 && !(v10x->struct_byte_0xc_12_15.byte[0] & 0x20))
@@ -19406,10 +19406,10 @@ void sub_1FAA0(type_event_0x6E8E* a1x)//200aa0
 					v19x = 0;
 					for (jz = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[36 / 4]; jz > x_DWORD_EA3E4[0]; jz = jz->next_0)
 					{
-						v20 = (signed __int16)(jz->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+						v20 = (signed __int16)((int16_t)jz->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 						v21 = v20 * v20
-							+ (signed __int16)(jz->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-							* (signed __int16)(jz->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+							+ (signed __int16)((int16_t)jz->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+							* (signed __int16)((int16_t)jz->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 						if (v21 <= v33 && v21 < v18)
 						{
 							v19x = jz;
@@ -19437,9 +19437,9 @@ void sub_1FAA0(type_event_0x6E8E* a1x)//200aa0
 				{
 					if (!kx->word_0x32_50 && kx != a1x)
 					{
-						v25 = (signed __int16)(kx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
-						v36 = (signed __int16)(kx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-							* (signed __int16)(kx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
+						v25 = (signed __int16)((int16_t)kx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
+						v36 = (signed __int16)((int16_t)kx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+							* (signed __int16)((int16_t)kx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
 							+ v25 * v25;
 						//v36 = jx;
 						if (v36 <= v30)
@@ -19716,10 +19716,10 @@ void sub_203D0(type_event_0x6E8E* a1x)//2013d0
 			{
 				if (jx->subtype_0x40_64 == 2 && jx->word_0x1A_26 != a1x->word_0x1A_26)
 				{
-					v11 = (signed __int16)(jx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+					v11 = (signed __int16)((int16_t)jx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 					v12 = v11 * v11
-						+ (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-						* (signed __int16)(jx->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+						+ (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+						* (signed __int16)((int16_t)jx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 					if (v12 < v8)
 					{
 						v9x = jx;
@@ -19763,9 +19763,9 @@ void sub_203D0(type_event_0x6E8E* a1x)//2013d0
 					{
 						if (v17x->word_0x1A_26 != a1x->word_0x1A_26)
 						{
-							v18 = (signed __int16)(v17x->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+							v18 = (signed __int16)((int16_t)v17x->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 							v19 = v18 * v18;
-							v20 = (signed __int16)(v17x->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+							v20 = (signed __int16)((int16_t)v17x->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 							v43 = v20 * v20 + v19;
 							if (v43 <= v42 && !(v17x->struct_byte_0xc_12_15.byte[0] & 0x20))
 							{
@@ -19820,10 +19820,10 @@ void sub_203D0(type_event_0x6E8E* a1x)//2013d0
 						v27x = 0;
 						for (kx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[16 / 4]; kx > x_DWORD_EA3E4[0]; kx = kx->next_0)
 						{
-							v29 = (signed __int16)(kx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+							v29 = (signed __int16)((int16_t)kx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 							v30 = v29 * v29
-								+ (signed __int16)(kx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-								* (signed __int16)(kx->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+								+ (signed __int16)((int16_t)kx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+								* (signed __int16)((int16_t)kx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 							if (v30 <= v45 && v30 < v26)
 							{
 								v27x = kx;
@@ -19838,10 +19838,10 @@ void sub_203D0(type_event_0x6E8E* a1x)//2013d0
 						v27x = 0;
 						for (lx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[48 / 4]; lx > x_DWORD_EA3E4[0]; lx = lx->next_0)
 						{
-							v33 = (signed __int16)(lx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+							v33 = (signed __int16)((int16_t)lx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 							v34 = v33 * v33
-								+ (signed __int16)(lx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-								* (signed __int16)(lx->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+								+ (signed __int16)((int16_t)lx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+								* (signed __int16)((int16_t)lx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 							if (v34 <= v46 && v34 < v31)
 							{
 								v27x = lx;
@@ -19858,10 +19858,10 @@ void sub_203D0(type_event_0x6E8E* a1x)//2013d0
 						v27x = 0;
 						for (mx = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[52 / 4]; mx > x_DWORD_EA3E4[0]; mx = mx->next_0)
 						{
-							v37 = (signed __int16)(mx->axis_0x4C_76.x - a1x->axis_0x4C_76.x);
+							v37 = (signed __int16)((int16_t)mx->axis_0x4C_76.x - (int16_t)a1x->axis_0x4C_76.x);
 							v38 = v37 * v37
-								+ (signed __int16)(mx->axis_0x4C_76.y - a1x->axis_0x4C_76.y)
-								* (signed __int16)(mx->axis_0x4C_76.y - a1x->axis_0x4C_76.y);
+								+ (signed __int16)((int16_t)mx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y)
+								* (signed __int16)((int16_t)mx->axis_0x4C_76.y - (int16_t)a1x->axis_0x4C_76.y);
 							if (v38 <= v40 && v38 < v35)
 							{
 								v27x = mx;
